feat: add deprecated shims to keep b/w compatibility
