+++ apps/automated/src/main.ts
@@ -5,136 +5,6 @@ if (typeof NSDate !== 'undefined') {
 	start = java.lang.System.currentTimeMillis();
 }
 
-// Utils.android.dismissSoftInput();
-// Utils.ad.dismissSoftInput();
-
-// Utils.ios.animateWithSpring();
-
-// Utils.android.isRealDevice()
-
-// isAndroid
-// const {
-// 	GC,
-// 	SDK_VERSION,
-// 	RESOURCE_PREFIX,
-// 	FILE_PREFIX,
-// 	isFontIconURI,
-// 	isDataURI,
-// 	isFileOrResourcePath,
-// 	getFileExtension,
-// 	executeOnMainThread,
-// 	executeOnUIThread,
-// 	mainThreadify,
-// 	isMainThread,
-// 	dispatchToMainThread,
-// 	queueMacrotask,
-// 	queueGC,
-// 	debounce,
-// 	throttle,
-// 	releaseNativeObject,
-// 	convertString,
-// 	escapeRegexSymbols,
-
-// 	getModuleName,
-// 	openFile,
-// 	openUrl,
-// 	isRealDevice,
-
-// 	layout,
-// 	android,
-// 	// legacy (a lot of plugins use the shorthand "ad" Utils.ad instead of Utils.android)
-// 	ad,
-// 	ios,
-// 	dataSerialize,
-// 	dataDeserialize,
-// 	numberHasDecimals,
-// 	numberIs64Bit,
-// 	setTimeout,
-// 	setInterval,
-// 	clearInterval,
-// 	clearTimeout,
-// 	Source,
-// 	ClassInfo,
-// 	getClass,
-// 	getBaseClasses,
-// 	getClassInfo,
-// 	isBoolean,
-// 	isDefined,
-// 	isFunction,
-// 	isNullOrUndefined,
-// 	isNumber,
-// 	isObject,
-// 	isString,
-// 	isUndefined,
-// 	toUIString,
-// 	verifyCallback,
-// 	dismissSoftInput,
-// 	dismissKeyboard,
-// 	copyToClipboard,
-// 	isEmoji,
-// 	getDurationWithDampingFromSpring,
-
-// 	// wtf
-// } = Utils;
-
-// const {
-// 	launchEvent,
-// 	displayedEvent,
-// 	uncaughtErrorEvent,
-// 	discardedErrorEvent,
-// 	suspendEvent,
-// 	resumeEvent,
-// 	exitEvent,
-// 	foregroundEvent,
-// 	backgroundEvent,
-// 	lowMemoryEvent,
-// 	orientationChangedEvent,
-// 	systemAppearanceChangedEvent,
-// 	systemAppearanceChanged,
-// 	fontScaleChangedEvent,
-// 	setMaxRefreshRate,
-
-// 	getMainEntry,
-// 	getRootView,
-// 	resetRootView,
-// 	getResources,
-// 	setResources,
-// 	setCssFileName,
-// 	getCssFileName,
-// 	loadAppCss,
-// 	addCss,
-// 	on,
-// 	off,
-// 	notify,
-// 	hasListeners,
-// 	run,
-// 	orientation,
-// 	getNativeApplication,
-// 	hasLaunched,
-// 	systemAppearance,
-// 	setAutoSystemAppearanceChanged,
-// 	android: appAndroid,
-// 	inBackground,
-// 	suspended,
-// 	ios: iosApp,
-// 	// get android() {
-// 	// 	ensureNativeApplication();
-// 	// 	return appAndroid;
-// 	// },
-// 	// get ios() {
-// 	// 	ensureNativeApplication();
-// 	// 	return iosApp;
-// 	// },
-// 	// get suspended() {
-// 	// 	return suspended;
-// 	// },
-// 	// get inBackground() {
-// 	// 	return inBackground;
-// 	// },
-// } = Application;
-
-// Application.suspendEvent
-
 import {
 	Application,
 	ApplicationEventData,
@@ -148,10 +18,7 @@ import {
 	AndroidActivityBackPressedEventData,
 } from '@nativescript/core';
 
-// import { Application } from "@nativescript/core/application-new";
-
-// Application.AndroidApplication
-// Application2.AndroidApplication
+// import * as Application from "@nativescript/core/application";
 
 if (Application.ios) {
 	// Observe application notifications.
