+++ packages/core/application/application-shims.ts
@@ -1,0 +1,393 @@
+import { Application } from './application';
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.addCss()
+ * ```
+ */
+export const addCss = Application.addCss.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.getCssFileName()
+ * ```
+ */
+export const getCssFileName = Application.getCssFileName.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.getMainEntry()
+ * ```
+ */
+export const getMainEntry = Application.getMainEntry.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.getNativeApplication()
+ * ```
+ */
+export const getNativeApplication = Application.getNativeApplication.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.getResources()
+ * ```
+ */
+export const getResources = Application.getResources.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.getRootView()
+ * ```
+ */
+export const getRootView = Application.getRootView.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.hasLaunched()
+ * ```
+ */
+export const hasLaunched = Application.hasLaunched.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.hasListeners()
+ * ```
+ */
+export const hasListeners = Application.hasListeners.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.loadAppCss()
+ * ```
+ */
+export const loadAppCss = Application.loadAppCss.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.notify()
+ * ```
+ */
+export const notify = Application.notify.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.off()
+ * ```
+ */
+export const off = Application.off.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.on()
+ * ```
+ */
+export const on = Application.on.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.orientation
+ * ```
+ */
+export const orientation = () => Application.orientation;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.resetRootView()
+ * ```
+ */
+export const resetRootView = Application.resetRootView.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.resetRootView()
+ * ```
+ */
+export const _resetRootView = Application.resetRootView.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.run()
+ * ```
+ */
+export const run = Application.run.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.setAutoSystemAppearanceChanged()
+ * ```
+ */
+export const setAutoSystemAppearanceChanged =
+	Application.setAutoSystemAppearanceChanged.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.setCssFileName()
+ * ```
+ */
+export const setCssFileName = Application.setCssFileName.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.setMaxRefreshRate()
+ * ```
+ */
+export const setMaxRefreshRate = Application.setMaxRefreshRate.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.setResources()
+ * ```
+ */
+export const setResources = Application.setResources.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.systemAppearance
+ * ```
+ */
+export const systemAppearance = () => Application.systemAppearance;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.systemAppearanceChanged()
+ * ```
+ */
+export const systemAppearanceChanged =
+	Application.systemAppearanceChanged.bind(Application);
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.discardedErrorEvent
+ * ```
+ */
+export const discardedErrorEvent = Application.discardedErrorEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.uncaughtErrorEvent
+ * ```
+ */
+export const uncaughtErrorEvent = Application.uncaughtErrorEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.displayedEvent
+ * ```
+ */
+export const displayedEvent = Application.displayedEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.exitEvent
+ * ```
+ */
+export const exitEvent = Application.exitEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.fontScaleChangedEvent
+ * ```
+ */
+export const fontScaleChangedEvent = Application.fontScaleChangedEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.launchEvent
+ * ```
+ */
+export const launchEvent = Application.launchEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.lowMemoryEvent
+ * ```
+ */
+export const lowMemoryEvent = Application.lowMemoryEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.resumeEvent
+ * ```
+ */
+export const resumeEvent = Application.resumeEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.orientationChangedEvent
+ * ```
+ */
+export const orientationChangedEvent = Application.orientationChangedEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.suspendEvent
+ * ```
+ */
+export const suspendEvent = Application.suspendEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.systemAppearanceChangedEvent
+ * ```
+ */
+export const systemAppearanceChangedEvent = Application.systemAppearanceChangedEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.backgroundEvent
+ * ```
+ */
+export const backgroundEvent = Application.backgroundEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.foregroundEvent
+ * ```
+ */
+export const foregroundEvent = Application.foregroundEvent;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.inBackground
+ * ```
+ */
+export const inBackground = Application.inBackground;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.suspended
+ * ```
+ */
+export const suspended = Application.suspended;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.android
+ * ```
+ */
+export const android = Application.android;
+
+/**
+ * @deprecated Deep imports into the Application module are depreacted and will be removed in a future release.
+ * Use the `Application` class imported from "@nativescript/core" instead:
+ * ```ts
+ * import { Application } from "@nativescript/core";
+ * Application.ios
+ * ```
+ */
+export const ios = Application.ios;
